Moved zod generator to dedicated export
